requirements.txt
Flask==3.1.1
numpy==2.3.0
pandas==2.3.0
Pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
